refactor(disposals): migrate type definitions to types/disposals.ts
