+++ src/types/disposals.ts
@@ -34,3 +34,32 @@ export type DisposalHistoryDetails = {
   status: string;
   documentUrls: string[];
 };
+
+export interface PendingDisposalRow {
+  id: number;
+  assetId: string;
+  assetTag: string;
+  assetName: string | null;
+  flaggedBy: string;
+  reason: string;
+  requestedAt: Date;
+}
+
+export interface HistoryDisposalRow {
+  id: number;
+  assetId: string;
+  assetTag: string;
+  category: string;
+  reason: string;
+  flaggedBy: string;
+  disposedBy: string | null;
+  disposalDate: Date | null;
+  status: string;
+  documentUrls: string[];
+}
+
+export type SelectedAssetLite = {
+  id: string;
+  assetTag: string;
+  assetName: string;
+};
